Adding first pass at script injection.
Adding two new options that if set, will attempt to scan the first 1k of
body data for the either the header tag or the html5 doctype tag (html,
head, and body are optional in html5) and inject a script tag
immediately after.  The first option is ScriptSource, which will inject
something like the following <script src="<ScriptSource>"
type="text/javascript"></script>.  The second option allows a file on
disk to provide the script content to be inject (nice because you can
avoid extra network requests).  Issues I suspect you will run into if
you attempt to use this are building on mac (I attempted and it didn't
work) and my use of the regex.h library to search for the head tag and
doctype.  Depending on where you build, the head tag regular expression…
<head[^>]*> may give you some trouble.  I developed and tested on Ubuntu
10.04.    Best, paulcnichols.